// sbh-fe-ts/src/enums/apiRoutes.enum.ts
export enum ApiRoutesEnum {
    TEST = 'http://a0540172.xsph.ru/api/testconnection',
    GET_MATH_FORMULAS = 'http://a0540172.xsph.ru/api/Math?name',
    GET_MATH_FORMULAS_BY_ID = 'http://a0540172.xsph.ru/api/Math/',
    GET_GEOMETRY_FORMULAS = 'http://a0540172.xsph.ru/api/Geometry?name',
    GET_GEOMETRY_FORMULAS_BY_ID = 'http://a0540172.xsph.ru/api/Geometry/',
    GET_PHYSICS_FORMULAS = 'http://a0540172.xsph.ru/api/Physics?name',
    GET_PHYSICS_FORMULAS_BY_ID = 'http://a0540172.xsph.ru/api/Physics/',
    GET_INFORMATICS_FORMULAS = 'http://a0540172.xsph.ru/api/Informatics?name',
    GET_INFORMATICS_FORMULAS_BY_ID = 'http://a0540172.xsph.ru/api/Informatics/'
}